requirements.txt
reedsolo==1.5.4
pycryptodome==3.9.8